requirements.txt
fastapi
supabase
requests
python-dotenv
uvicorn
gunicorn
python-multipart
